move metafields container to containers folder
