+++ lib/jekyll/admin/public/docs/containers.md
@@ -23,6 +23,27 @@ Container for displaying header which includes title and homepage link.
 }
 ```
 
+## MetaFields
+Main container for metafields.
+
+All of the prop types below passed down to the children;
+
+** PropTypes **
+``` javascript
+{
+  meta: Object, // passed from ContentEdit
+  metadata: Object, // passed from Redux store
+  key_prefix: String,
+  storeContentFields: Function,
+  addField: Function,
+  removeField: Function,
+  updateFieldKey: Function,
+  updateFieldValue: Function,
+  moveArrayItem: Function,
+  convertField: Function
+}
+```
+
 ## Views
 Contains all of the views linked with the routes.
