--- lib/jekyll/admin/public/src/components/metadata/tests/fixtures/index.js
@@ -1,27 +1,0 @@
-export const meta = {
-  layout: "post",
-  categories: "gsoc",
-  students: [
-    "GSoC Students",
-    {
-      name: {
-        first: "Mert",
-        last: "Kahyaoğlu"
-      },
-      email: [
-        "mertkahyaoglu93@gmail.com",
-        "test@gmail.com"
-      ],
-      username: "mertkahyaoglu"
-    },
-    {
-      name: {
-        first: "Ankur",
-        last: "Singh"
-      },
-      email: "ankur13019@iiitd.ac.in",
-      username: "rush-skills"
-    }
-  ],
-  mentors: ["Ben Balter", "Jurgen Leschner", "Parker Moore"]
-};
