+++ lib/jekyll/admin/public/src/containers/MetaFields.js
@@ -4,19 +4,23 @@ import { bindActionCreators } from 'redux';
 import _ from 'underscore';
 
 // Components
-import MetaField from './MetaField';
+import MetaField from '../components/metadata/MetaField';
 
 // Actions
 import {
-  addField, removeField, setupMetadata, updateFieldKey, updateFieldValue,
+  storeContentFields, addField, removeField, updateFieldKey, updateFieldValue,
   moveArrayItem, convertField
-} from '../../actions/metadata';
+} from '../actions/metadata';
 
 export class MetaFields extends Component {
 
-  componentWillMount() {
-    const { setupMetadata, meta } = this.props;
-    setupMetadata(meta);
+  componentDidMount() {
+    const { storeContentFields, content } = this.props;
+    storeContentFields(content);
+  }
+
+  shouldComponentUpdate(nextProps) {
+    return nextProps.metadata !== this.props.metadata;
   }
 
   render() {
@@ -59,10 +63,10 @@ export class MetaFields extends Component {
 }
 
 MetaFields.propTypes = {
-  meta: PropTypes.object.isRequired,
+  content: PropTypes.object.isRequired,
   metadata: PropTypes.object.isRequired,
   key_prefix: PropTypes.string.isRequired,
-  setupMetadata: PropTypes.func.isRequired,
+  storeContentFields: PropTypes.func.isRequired,
   addField: PropTypes.func.isRequired,
   removeField: PropTypes.func.isRequired,
   updateFieldKey: PropTypes.func.isRequired,
@@ -81,7 +85,7 @@ function mapStateToProps(state) {
 
 function mapDispatchToProps(dispatch) {
   return bindActionCreators({
-    setupMetadata,
+    storeContentFields,
     addField,
     removeField,
     updateFieldKey,
