+++ lib/jekyll/admin/public/src/containers/tests/fixtures/index.js
@@ -13,3 +13,36 @@ export const collections = [
   {title: "Posts", path: "/posts" },
   {title: "Movies", path: "/movies"}
 ];
+
+export const content = {
+  body: "Body",
+  meta: {
+    title: "GSoC",
+    layout: "post",
+    path: "gsoc.md",
+    categories: "gsoc",
+    students: [
+      "GSoC Students",
+      {
+        name: {
+          first: "Mert",
+          last: "Kahyaoğlu"
+        },
+        email: [
+          "mertkahyaoglu93@gmail.com",
+          "test@gmail.com"
+        ],
+        username: "mertkahyaoglu"
+      },
+      {
+        name: {
+          first: "Ankur",
+          last: "Singh"
+        },
+        email: "ankur13019@iiitd.ac.in",
+        username: "rush-skills"
+      }
+    ],
+    mentors: ["Ben Balter", "Jurgen Leschner", "Parker Moore"]
+  }
+};
