+++ lib/jekyll/admin/public/src/containers/tests/metafields.spec.js
@@ -4,14 +4,14 @@ import expect from 'expect';
 import { mount } from 'enzyme';
 import _ from 'underscore';
 
-import { MetaFields } from '../index';
-import MetaField from '../MetaField';
+import { MetaFields } from '../MetaFields';
+import MetaField from '../../components/metadata/MetaField';
 
-import { meta } from './fixtures';
+import { content } from './fixtures';
 
 function setup() {
   const actions = {
-    setupMetadata: expect.createSpy(),
+    storeContentFields: expect.createSpy(),
     addField: expect.createSpy(),
     removeField: expect.createSpy(),
     updateFieldKey: expect.createSpy(),
@@ -22,8 +22,8 @@ function setup() {
 
   let component = mount(
     <MetaFields
-      meta={meta}
-      metadata={meta}
+      content={content}
+      metadata={content.meta}
       key_prefix=""
       {...actions} />
   );
@@ -40,15 +40,13 @@ describe('Components::MetaFields', () => {
   it('should render MetaFields correctly', () => {
     const { component, metafields } = setup();
     expect(component.prop('key_prefix')).toBe('');
-    expect(component.prop('metadata')).toEqual(meta);
-    expect(metafields.length).toBe(_.keys(meta).length);
+    expect(component.prop('metadata')).toEqual(content.meta);
+    expect(metafields.length).toBe(_.keys(content.meta).length);
   });
-
-  it('should call setupMetadata before mount', () => {
+  it('should call storeContentFields before mount', () => {
     const { actions } = setup();
-    expect(actions.setupMetadata).toHaveBeenCalled();
+    expect(actions.storeContentFields).toHaveBeenCalled();
   });
-
   it('should call addField when the button is clicked', () => {
     const { actions, addFieldButton } = setup();
     addFieldButton.simulate('click');
